+++ src/_playground/documentation/Properties/Properties.js
@@ -16,13 +16,27 @@ export const Properties = ({ sourceModulePath }) => {
             <h2>Properties</h2>
             {componentNames.map((name, index) => {
                 const component = sourceModule[name];
+                const subcomponentNames = Object.keys(component)
+                    .sort()
+                    .filter(subName => component[subName] && component[subName].displayName);
                 return (
                     <React.Fragment key={index}>
-                        <h3 className='header'>{name}</h3>
                         <PropertyTable
                             defaultProps={component.defaultProps}
                             propDescriptions={component.propDescriptions}
-                            propTypes={component.propTypes} />
+                            propTypes={component.propTypes}
+                            title={name} />
+                        {subcomponentNames.map((subName, subIndex) => {
+                            const subcomponent = component[subName];
+                            return (
+                                <PropertyTable
+                                    defaultProps={subcomponent.defaultProps}
+                                    key={subIndex}
+                                    propDescriptions={subcomponent.propDescriptions}
+                                    propTypes={subcomponent.propTypes}
+                                    title={`${name}.${subName}`} />
+                            );
+                        })}
                     </React.Fragment>
                 );
             })}
@@ -36,7 +50,7 @@ Properties.propTypes = {
 
 
 
-const PropertyTable = ({ propTypes, defaultProps, propDescriptions }) => {
+const PropertyTable = ({ title, propTypes, defaultProps, propDescriptions }) => {
     if (!propTypes) {
         return (
             <em>This component has no defined properties.</em>
@@ -71,20 +85,24 @@ const PropertyTable = ({ propTypes, defaultProps, propDescriptions }) => {
     });
 
     return (
-        <Table
-            className='property-table'
-            headers={
-                [
-                    'Name',
-                    'Type',
-                    'Default',
-                    'Description'
-                ]}
-            tableData={data} />
+        <React.Fragment>
+            <h3 className='header'>{title}</h3>
+            <Table
+                className='property-table'
+                headers={
+                    [
+                        'Name',
+                        'Type',
+                        'Default',
+                        'Description'
+                    ]}
+                tableData={data} />
+        </React.Fragment>
     );
 };
 
 PropertyTable.propTypes = {
+    title: PropTypes.string.isRequired,
     defaultProps: PropTypes.object,
     propDescriptions: PropTypes.object,
     propTypes: PropTypes.object
@@ -94,55 +112,62 @@ PropertyTable.propTypes = {
 
 const PropertyType = ({ prop }) => {
     const typeChecker = prop.typeChecker;
-    let details;
 
     switch (prop.typeName) {
         case 'arrayOf':
         case 'objectOf':
-            details = (
-                <div>({typeChecker.typeName})</div>
+            return (
+                <React.Fragment>
+                    <div>{prop.typeName}</div>
+                    {typeChecker.typeChecker ? (
+                        <PropertyType prop={typeChecker} />
+                    ) : (
+                        <div>{`(${typeChecker.typeName})`}</div>
+                    )}
+                </React.Fragment>
             );
-            break;
         case 'instanceOf':
-            details = (
-                <div>({typeChecker.name || typeChecker.displayName})</div>
+            return (
+                <React.Fragment>
+                    <div>{prop.typeName}</div>
+                    {typeChecker.typeChecker ? (
+                        <PropertyType prop={typeChecker} />
+                    ) : (
+                        <div>{`(${typeChecker.name || typeChecker.displayName})`}</div>
+                    )}
+                </React.Fragment>
             );
-            break;
         case 'oneOf':
-            details = (
-                <div>({typeChecker.join(', ')})</div>
+            return (
+                <React.Fragment>
+                    <div>{prop.typeName}</div>
+                    <div>{`(${typeChecker.join(', ')})`}</div>
+                </React.Fragment>
             );
-            break;
         case 'oneOfType':
-            let types = [];
-            typeChecker.forEach(t => {
-                types.push(t.typeName);
-            });
-            details = (
-                <div>({types.join(', ')})</div>
+            const types = typeChecker
+                .map((t, i) =>
+                    t.typeChecker ? <PropertyType key={i} prop={t} /> : <span key={i}>{`(${t.typeName})`}</span>
+                ).reduce((prev, curr) => [prev, ', ', curr]);
+            return (
+                <React.Fragment>
+                    <div>{prop.typeName}</div>
+                    [{types}]
+                </React.Fragment>
             );
-            break;
         case 'shape':
-            details = '';
-            break;
+            return <div>{prop.typeName} ({`\{${Object.keys(prop.typeChecker).sort().join(', ')}\}`})</div>;
         case 'range':
-            let values = [];
-            Object.keys(typeChecker).forEach(key => {
-                values.push(`${key}: ${typeChecker[key]}`);
-            });
-            details = (
-                <div>({values.join('; ')})</div>
+            const values = Object.keys(typeChecker).map(key => `${key}: ${typeChecker[key]}`);
+            return (
+                <React.Fragment>
+                    <div>{prop.typeName}</div>
+                    <div>{`(${values.join('; ')})`}</div>
+                </React.Fragment>
             );
-            break;
         default:
+            return <div>{prop.typeName}</div>;
     }
-
-    return (
-        <React.Fragment>
-            <div>{prop.typeName}</div>
-            <div>{details}</div>
-        </React.Fragment>
-    );
 };
 
 PropertyType.propTypes = {